+++ core/tauri/src/manager/webview.rs
@@ -89,6 +89,9 @@ pub struct WebviewManager<R: Runtime> {
   pub invoke_responder: Option<Arc<InvokeResponder<R>>>,
   /// The script that initializes the invoke system.
   pub invoke_initialization_script: String,
+
+  /// A runtime generated invoke key.
+  pub(crate) invoke_key: String,
 }
 
 impl<R: Runtime> fmt::Debug for WebviewManager<R> {
@@ -98,6 +101,7 @@ impl<R: Runtime> fmt::Debug for WebviewManager<R> {
         "invoke_initialization_script",
         &self.invoke_initialization_script,
       )
+      .field("invoke_key", &self.invoke_key)
       .finish()
   }
 }
@@ -133,7 +137,7 @@ impl<R: Runtime> WebviewManager<R> {
     let app_manager = manager.manager();
 
     let is_init_global = app_manager.config.app.with_global_tauri;
-    let plugin_init = app_manager
+    let plugin_init_scripts = app_manager
       .plugins
       .lock()
       .expect("poisoned plugin store")
@@ -199,9 +203,12 @@ impl<R: Runtime> WebviewManager<R> {
         app_manager,
         &ipc_init.into_string(),
         &pattern_init.into_string(),
-        &plugin_init,
         is_init_global,
       )?);
+
+    for plugin_init_script in plugin_init_scripts {
+      webview_attributes = webview_attributes.initialization_script(&plugin_init_script);
+    }
 
     #[cfg(feature = "isolation")]
     if let crate::Pattern::Isolation { schema, .. } = &*app_manager.pattern {
@@ -246,16 +253,18 @@ impl<R: Runtime> WebviewManager<R> {
       && window_url.scheme() != "https"
     {
       format!("http://{}.localhost", window_url.scheme())
-    } else {
+    } else if let Some(host) = window_url.host() {
       format!(
         "{}://{}{}",
         window_url.scheme(),
-        window_url.host().unwrap(),
+        host,
         window_url
           .port()
           .map(|p| format!(":{p}"))
           .unwrap_or_default()
       )
+    } else {
+      "null".into()
     };
 
     if !registered_scheme_protocols.contains(&"tauri".into()) {
@@ -272,8 +281,7 @@ impl<R: Runtime> WebviewManager<R> {
     }
 
     if !registered_scheme_protocols.contains(&"ipc".into()) {
-      let protocol =
-        crate::ipc::protocol::get(manager.manager_owned().clone(), pending.label.clone());
+      let protocol = crate::ipc::protocol::get(manager.manager_owned(), pending.label.clone());
       pending.register_uri_scheme_protocol("ipc", move |request, responder| {
         protocol(request, UriSchemeResponder(responder))
       });
@@ -345,7 +353,6 @@ impl<R: Runtime> WebviewManager<R> {
     app_manager: &AppManager<R>,
     ipc_script: &str,
     pattern_script: &str,
-    plugin_initialization_script: &str,
     with_global_tauri: bool,
   ) -> crate::Result<String> {
     #[derive(Template)]
@@ -362,8 +369,6 @@ impl<R: Runtime> WebviewManager<R> {
       #[raw]
       event_initialization_script: &'a str,
       #[raw]
-      plugin_initialization_script: &'a str,
-      #[raw]
       freeze_prototype: &'a str,
     }
 
@@ -371,6 +376,7 @@ impl<R: Runtime> WebviewManager<R> {
     #[default_template("../../scripts/core.js")]
     struct CoreJavascript<'a> {
       os_name: &'a str,
+      invoke_key: &'a str,
     }
 
     let bundle_script = if with_global_tauri {
@@ -391,6 +397,7 @@ impl<R: Runtime> WebviewManager<R> {
       bundle_script,
       core_script: &CoreJavascript {
         os_name: std::env::consts::OS,
+        invoke_key: self.invoke_key(),
       }
       .render_default(&Default::default())?
       .into_string(),
@@ -398,7 +405,6 @@ impl<R: Runtime> WebviewManager<R> {
         app_manager.listeners().function_name(),
         app_manager.listeners().listeners_object_name(),
       ),
-      plugin_initialization_script,
       freeze_prototype,
     }
     .render_default(&Default::default())
@@ -614,7 +620,7 @@ impl<R: Runtime> WebviewManager<R> {
     }
 
     // let plugins know that a new webview has been added to the manager
-    let manager = webview.manager_owned().clone();
+    let manager = webview.manager_owned();
     let webview_ = webview.clone();
     // run on main thread so the plugin store doesn't dead lock with the event loop handler in App
     let _ = webview.run_on_main_thread(move || {
@@ -660,6 +666,10 @@ impl<R: Runtime> WebviewManager<R> {
 
   pub fn labels(&self) -> HashSet<String> {
     self.webviews_lock().keys().cloned().collect()
+  }
+
+  pub(crate) fn invoke_key(&self) -> &str {
+    &self.invoke_key
   }
 }