+++ core/tauri-utils/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## \[2.0.0-beta.16]
+
+### Bug Fixes
+
+- [`be95d8d37`](https://www.github.com/tauri-apps/tauri/commit/be95d8d37c4b1a420c0d28a83b7efa40ab0b0ab5)([#9782](https://www.github.com/tauri-apps/tauri/pull/9782)) Fixes the `ToTokens` implementation for `Capability`.
+
+## \[2.0.0-beta.15]
+
+### Bug Fixes
+
+- [`a5205f179`](https://www.github.com/tauri-apps/tauri/commit/a5205f179e577cce5c05b710b597da8f4e1d0780)([#9691](https://www.github.com/tauri-apps/tauri/pull/9691)) Fixes the ToTokens implementation of the window configuration `proxy_url` field.
+
+## \[2.0.0-beta.14]
+
+### Bug Fixes
+
+- [`3fbc1703f`](https://www.github.com/tauri-apps/tauri/commit/3fbc1703f107dfb2c5a75e848805dcf60d449eb1)([#9676](https://www.github.com/tauri-apps/tauri/pull/9676)) Fixes `schemars` compilation issue.
+
 ## \[2.0.0-beta.13]
 
 ### Bug Fixes